requirements.txt
Django
psycopg2-binary
python-dotenv
django-crispy-forms
crispy_bootstrap5
requests>=2.31.0 
cryptography
PyJWT
pycparser==2.22
django-allauth
